Add PDA light selection support
In previous version, a virtual button can be used to cycle over light
color. With this version, it supports selecting a specific color from
Web UI and HomeKit.

For HomeKit, the device is implemented as a DIMMER instead a switch and
will require AqualinkD HomeKit to be updated.

Though, the virtual cycle buton can be use with HomeKit if you don't
want a dimmer style button.

Signed-off-by: locnho <locnhinho@gmail.com>